# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/files": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Список файлов",
                "operationId": "5baabf60af4423ca7af8c73f4f8cf994",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "table_name": {
                                                        "type": "string",
                                                        "example": "table_name"
                                                    },
                                                    "row_id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "file_id": {
                                                        "type": "string"
                                                    },
                                                    "filename": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Импорт данных файла",
                "operationId": "9a2bf45b41e0ce24050fa66025747fc6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "table_name",
                                    "row_id",
                                    "filepath"
                                ],
                                "properties": {
                                    "table_name": {
                                        "type": "string",
                                        "example": "table_name"
                                    },
                                    "row_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "file_id": {
                                        "type": "string"
                                    },
                                    "filename": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/files/{id}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Получить файл по id",
                "operationId": "438d254fdec671d1d74fc30d64a80028",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "table_name": {
                                                    "type": "string"
                                                },
                                                "row_id": {
                                                    "type": "integer"
                                                },
                                                "file_id": {
                                                    "type": "string"
                                                },
                                                "filename": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Files"
                ],
                "summary": "Удалить данные о файле",
                "operationId": "dfd9c6f79f5a54b06a9491ad78e3f045",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Test endpoint",
                "operationId": "7a09caa81b6dfd4c5087d06d7689eaa8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Files"
        },
        {
            "name": "Test"
        }
    ]
}